+++ test/support.js
@@ -1,7 +1,6 @@
 var fs = require('fs')
 var test = require('tape')
 var mingo = require('../es5')
-var _ = mingo._internal()
 
 exports.personData = JSON.parse(fs.readFileSync(__dirname + '/data/person.json'))
 exports.simpleGradesData = JSON.parse(fs.readFileSync(__dirname + '/data/grades_simple.json'))
@@ -17,10 +16,19 @@ exports.groupByObjectsData = [
   {'date_buckets': {'date': '2015-04-29T00:17:03.107Z', 'day': 28, 'hour': 18, 'minute': 17, 'sec': 3, 'hour_minute': '18:17'}, 'Keyword ID': 'sr3_irU8fFk0', 'Creative ID': '6074827289', 'Keyword': 'unclog bathtub drain', 'Match Type': 'bp', 'Device': 'c', 'Conversions': [1, 0, 0, 1, 0, 0, 0, 0, 0], 'Revenues': [5, 0, 0, 5, 0, 0, 0, 0, 0], 'account_id': 'baron'}
 ]
 
+function computeValue(obj, input, field) {
+  var expr = {}
+  expr[field] = input
+  var result = mingo.aggregate([ obj ], [ { $addFields: {test: expr} }, { $project: { test: 1} } ])[0]
+  return result.test
+}
+
 exports.runTest = function (description, suite) {
-  _.each(suite, function (examples, operator) {
+  Object.entries(suite).forEach(function (arr) {
+    var operator = arr[0]
+    var examples = arr[1]
     test(description + ': ' + operator, function (t) {
-      _.each(examples, function (val) {
+      examples.forEach(function (val) {
         var input = val[0]
         var expected = val[1]
         var ctx = val[2] || { err: false }
@@ -28,8 +36,8 @@ exports.runTest = function (description, suite) {
 
         var field = operator
         // use the operator as field if not present in input
-        if (_.isObject(input)) {
-          field = _.keys(input).find((s) => s[0] === '$') || null
+        if (typeof input === 'object') {
+          field = Object.keys(input).find((s) => s[0] === '$') || null
           if (field === null) {
             field = operator
           } else {
@@ -38,9 +46,9 @@ exports.runTest = function (description, suite) {
         }
 
         if (ctx.err) {
-          t.throws(() => _.computeValue(obj, input, field),  JSON.stringify(input) + '\t=>\t' + expected)
+          t.throws(() => computeValue(obj, input, field), JSON.stringify(input) + '\t=>\t' + expected)
         } else {
-          var actual = _.computeValue(obj, input, field)
+          var actual = computeValue(obj, input, field)
           var message =  operator + ':\t' + JSON.stringify(input) + '\t=>\t' + JSON.stringify(expected)
           // NaNs don't compare
           if (actual !== actual && expected !== expected) actual = expected = 0
@@ -57,18 +65,16 @@ exports.runTest = function (description, suite) {
  */
 exports.runTestPipeline = function (description, suite) {
   test(description, function (t) {
-    _.each(suite, function (unitTest) {
+    suite.forEach(function (unitTest) {
       var pipeline = unitTest.query
       var input = unitTest.input
       var check = unitTest.check
-      var hash = _.hashCode(input)
       var actual = mingo.aggregate(input, pipeline)
-      if (_.isFunction(check)) {
+      if (typeof check === 'function') {
         check(actual, t)
       } else {
         t.deepEqual(actual, check, unitTest.message || "actual equals expected")
       }
-      _.assert(hash === _.hashCode(input), "input changed")
     })
     t.end()
   })